fix: three real failures surfaced by the post-merge CI run
Full suite now 486 passed, 0 failed against a real Postgres, reproduced
and fixed locally:

- conftest.py: COOKIE_SECURE now defaults to False in tests. The test
  client talks to the ASGI app over plain http://testserver, and
  httpx's cookie jar correctly refuses to resend a Secure-flagged
  cookie over a non-https base_url -- so every auth/session/remember-me
  test that round-trips a cookie across two calls (test_remember_me.py,
  test_sessions.py, test_oauth_session_metadata.py) failed with
  "Refresh token missing", which looked like an auth bug but was a
  transport/config mismatch. .env.example already documents the same
  COOKIE_SECURE=False for local dev for the identical reason.

- tests/ledger/test_migration.py: replaced relative `downgrade -1` with
  an explicit target resolved from the script graph. `-1` is only
  well-defined from a single-parent revision; the new merge migration
  (7bbdda89ceba, reconciling this branch's ledger/scoring chain with
  main's auth/lite-grading chain) has two parents, so "-1" is
  structurally ambiguous ("Ambiguous walk"). The new
  _one_step_back_from_head() resolves the real target from the current
  head's down_revision instead of hardcoding today's specific hash, so
  this keeps working after the next ordinary migration and after any
  future merge too.

- app/projects/schemas.py + engine_constraints.py: requested_amount
  changed from Decimal to int, matching loan_applications.requested_amount's
  actual Numeric(20,0) column (T1, HANDOFF-03 -- VND has no sub-unit).
  A Decimal here could round-trip through Postgres NUMERIC and come
  back as e.g. Decimal('8E+8') for a round amount like 800,000,000,
  since Postgres doesn't guarantee a non-exponential internal form and
  asyncpg preserves whatever form it sends -- a real formatting bug,
  not a test artifact. test_create_project_inline_loan.py's assertion
  was asserting the pre-T1 string form ("800000000.00"); updated to a
  plain JSON integer.

+++ app/projects/schemas.py
@@ -1,6 +1,5 @@
 from pydantic import BaseModel, field_validator, model_validator
 from datetime import date, datetime
-from decimal import Decimal
 from typing import Any
 from uuid import UUID
 
@@ -9,7 +8,14 @@ from app.uploads.schemas import LoanApplicationDocumentOut
 
 
 class LoanApplicationCreateInline(BaseModel):
-    requested_amount: Decimal
+    # VND has no sub-unit (T1, HANDOFF-03) and `loan_applications.requested_amount`
+    # is Numeric(20,0) at the DB layer -- `int` end to end, never Decimal. A
+    # Decimal here used to round-trip through Postgres NUMERIC and come back
+    # as e.g. Decimal('8E+8') for a round amount like 800,000,000 (Postgres
+    # doesn't guarantee a non-exponential internal form, and asyncpg preserves
+    # whatever form it sends), so the API could silently serialize a clean
+    # amount as scientific notation. `int` sidesteps the whole class of bug.
+    requested_amount: int
     # Loan term in months. Optional so an existing client that predates the
     # field keeps working; validated against the engine's
     # allowed_durations_months when present. NOTE: not persisted yet —
@@ -21,7 +27,7 @@ class LoanApplicationCreateInline(BaseModel):
 
     @field_validator("requested_amount")
     @classmethod
-    def _amount_within_engine_bounds(cls, value: Decimal) -> Decimal:
+    def _amount_within_engine_bounds(cls, value: int) -> int:
         return engine_constraints.validate_loan_size(value)
 
     @field_validator("duration_months")
@@ -88,7 +94,7 @@ class ProjectOut(BaseModel):
 class ProjectLoanApplicationOut(BaseModel):
     id: UUID
     project_id: UUID
-    requested_amount: Decimal
+    requested_amount: int
     purpose: str | None
     repayment_preference: str | None
     status: str
